minor reorganization of types
