+++ WebdriverIO/HW_19/src/data/types/product.types.ts
@@ -16,8 +16,3 @@ export enum MANUFACTURERS {
   AMAZON = 'Amazon',
   TESLA = 'Tesla'
 }
-
-export enum PRODUCT_TOAST_STATUSES {
-  DELETED = 'Product was successfully deleted',
-  CREATED = 'Product was successfully created'
-}
